pyproject.toml
[tool.poetry]
name = "pet-posts"
version = "0.1.0"
description = ""
authors = ["Dawn James <dawnjames2105@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.7"
python-dotenv = "^0.19.0"
black = "^21.9b0"

[tool.poetry.dev-dependencies]
pytest = "^6"
jurigged = "^0.3.4"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
